Add Storage and Command class diagrams
Add CalculateInputCommand sequence diagrams

Refactor and shift helper methods to helper class
