+++ src/main/java/seedu/mindmymoney/command/AddCommand.java
@@ -4,6 +4,7 @@ import seedu.mindmymoney.MindMyMoneyException;
 import seedu.mindmymoney.data.CreditCardList;
 import seedu.mindmymoney.data.ExpenditureList;
 import seedu.mindmymoney.data.IncomeList;
+import seedu.mindmymoney.helper.TimeFunctions;
 import seedu.mindmymoney.userfinancial.Expenditure;
 import seedu.mindmymoney.userfinancial.CreditCard;
 import seedu.mindmymoney.userfinancial.Income;
@@ -21,22 +22,16 @@ import static seedu.mindmymoney.constants.Flags.FLAG_OF_EXPENSES;
 import static seedu.mindmymoney.constants.Flags.FLAG_OF_INCOME;
 import static seedu.mindmymoney.constants.Flags.FLAG_OF_PAYMENT_METHOD;
 import static seedu.mindmymoney.constants.Flags.FLAG_OF_TIME;
-import static seedu.mindmymoney.constants.Indexes.INDEX_OF_FIRST_ITEM;
-import static seedu.mindmymoney.constants.Indexes.INDEX_OF_SECOND_ITEM;
-import static seedu.mindmymoney.constants.Indexes.INDEX_OF_THIRD_ITEM;
-import static seedu.mindmymoney.helper.AddCommandInputTests.isValidInput;
 import static seedu.mindmymoney.helper.AddCommandInputTests.testCashbackAmount;
 import static seedu.mindmymoney.helper.AddCommandInputTests.testCreditCardLimit;
 import static seedu.mindmymoney.helper.AddCommandInputTests.testCreditCardName;
-import static seedu.mindmymoney.helper.AddCommandInputTests.testDescription;
-import static seedu.mindmymoney.helper.AddCommandInputTests.testExpenditureAmount;
-import static seedu.mindmymoney.helper.AddCommandInputTests.testExpenditureCategory;
+import static seedu.mindmymoney.helper.AddCommandInputTests.testExpenditureParameters;
 import static seedu.mindmymoney.helper.AddCommandInputTests.testIncomeAmount;
 import static seedu.mindmymoney.helper.AddCommandInputTests.testIncomeCategory;
-import static seedu.mindmymoney.helper.AddCommandInputTests.testPaymentMethod;
 import static seedu.mindmymoney.helper.GeneralFunctions.capitalise;
 import static seedu.mindmymoney.helper.GeneralFunctions.parseInputWithCommandFlag;
 import static seedu.mindmymoney.helper.GeneralFunctions.formatFloat;
+import static seedu.mindmymoney.helper.TimeFunctions.checkAfterCurrentDate;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -49,7 +44,6 @@ public class AddCommand extends Command {
     public ExpenditureList expenditureList;
     public CreditCardList creditCardList;
     public IncomeList incomeList;
-    private static final int LEAP_YEAR_NUMBER = 4;
 
     public AddCommand(String addInput, User user) {
         this.addInput = addInput;
@@ -114,25 +108,6 @@ public class AddCommand extends Command {
     }
 
     /**
-     * Tests if the input parameters from the user are valid.
-     *
-     * @param paymentMethod The payment method used, either as cash or the credit card.
-     * @param inputCategory The category as indicated by the user.
-     * @param description The description of the expenditure.
-     * @param amountAsString Price of the expenditure.
-     * @param inputTime Date of when the expenditure was bought.
-     * @throws MindMyMoneyException when the parameters are invalid.
-     */
-    public void testExpenditureParameters(String paymentMethod, String inputCategory, String description,
-                String amountAsString, String inputTime) throws MindMyMoneyException {
-        testPaymentMethod(paymentMethod, creditCardList);
-        testExpenditureCategory(inputCategory);
-        testDescription(description);
-        testExpenditureAmount(amountAsString, paymentMethod, creditCardList);
-        checkValidDate(inputTime);
-    }
-
-    /**
      * Inserts an Expenditure object into user's list of expenditure(s).
      *
      * @throws MindMyMoneyException when inputs are invalid or flags are missing.
@@ -144,7 +119,7 @@ public class AddCommand extends Command {
         String amountAsString = parseInputWithCommandFlag(addInput, FLAG_OF_AMOUNT, FLAG_OF_TIME);
         String inputTime = parseInputWithCommandFlag(addInput, FLAG_OF_TIME, FLAG_END_VALUE);
 
-        testExpenditureParameters(paymentMethod, inputCategory, description, amountAsString, inputTime);
+        testExpenditureParameters(paymentMethod, inputCategory, description, amountAsString, inputTime, creditCardList);
 
         if (capitalise(paymentMethod).equals("Cash")) {
             paymentMethod = capitalise(paymentMethod);
@@ -171,44 +146,6 @@ public class AddCommand extends Command {
             System.out.printf(paymentMethod + " has a balance of $%.2f left%n", balanceLeft);
         }
         System.out.print(System.lineSeparator());
-    }
-
-    /**
-     * Checks if parsed date is after the current date.
-     *
-     * @param date date that is parsed in.
-     * @throws MindMyMoneyException throws an exception when the date parsed is after current date.
-     */
-    public static void checkAfterCurrentDate(LocalDate date) throws MindMyMoneyException {
-        LocalDate currentDate = LocalDate.now();
-        if (date.isAfter(currentDate)) {
-            throw new MindMyMoneyException("Please enter a valid date that is before today or today's date itself.");
-        }
-    }
-
-    /**
-     * Checks is parsed date is a valid date in the calendar.
-     *
-     * @param inputTime date that is parsed in.
-     * @throws MindMyMoneyException throws an exception when the date parsed is in not in the calendar.
-     */
-    public static void checkValidDate(String inputTime) throws MindMyMoneyException {
-        String[] date = inputTime.split("/");
-        String day = date[INDEX_OF_FIRST_ITEM];
-        int dayInInt = Integer.parseInt(day);
-        String month = date[INDEX_OF_SECOND_ITEM];
-        String year = date[INDEX_OF_THIRD_ITEM];
-        int yearInInt = Integer.parseInt(year);
-        if (!isValidInput(inputTime)) {
-            throw new MindMyMoneyException("Date has to be valid and in this format \"dd/mm/yyyy\"");
-        } else if (!(yearInInt % LEAP_YEAR_NUMBER == 0) && month.equals("02") && (dayInInt > 28)) {
-            throw new MindMyMoneyException(day + "/" + month + " is not a valid dd/mm in a non leap year!");
-        } else if ((yearInInt % LEAP_YEAR_NUMBER == 0) && month.equals("02") && (dayInInt > 29)) {
-            throw new MindMyMoneyException(day + "/" + month + " is not a valid dd/mm in a leap year!");
-        } else if ((month.equals("04") || month.equals("06") || month.equals("09") || month.equals("11"))
-                && dayInInt > 30) {
-            throw new MindMyMoneyException(day + "/" + month + " is not a valid dd/mm in this month!");
-        }
     }
 
     /**
