+++ src/main/java/seedu/mindmymoney/constants/CalculationConversion.java
@@ -3,6 +3,6 @@ package seedu.mindmymoney.constants;
 /**
  * Container for numbers used in calculations.
  */
-public class Calculations {
+public class CalculationConversion {
     public static final float FLOAT_TO_PERCENTAGE = (float) 0.01;
 }
