+++ src/main/java/seedu/mindmymoney/userfinancial/CreditCard.java
@@ -4,7 +4,7 @@ package seedu.mindmymoney.userfinancial;
 import seedu.mindmymoney.MindMyMoneyException;
 import seedu.mindmymoney.helper.PropertyList;
 
-import static seedu.mindmymoney.constants.Calculations.FLOAT_TO_PERCENTAGE;
+import static seedu.mindmymoney.constants.CalculationConversion.FLOAT_TO_PERCENTAGE;
 import static seedu.mindmymoney.helper.GeneralFunctions.formatFloat;
